pyproject.toml
[project]
name = "moxie-buildings"
version = "0.1.0"
requires-python = ">=3.12"
description = "Chicago rental market scraper and data aggregator"
dependencies = [
    "sqlalchemy==2.0.46",
    "alembic==1.18.4",
    "pydantic>=2.0",
    "gspread==6.2.1",
    "google-auth>=2.0",
    "python-dotenv>=1.0",
    "python-dateutil>=2.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.0",
    "crawl4ai>=0.8.0",
    "lxml>=5.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.11.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "PyJWT>=2.0",
    "pwdlib[argon2]>=0.2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.20",
]

[project.scripts]
sheets-sync = "moxie.sync.sheets:main"
export-platforms = "moxie.sync.export_platforms:main"
scrape = "moxie.scrape:main"
scrape-all = "moxie.scrape_all:main"
detect-platforms = "scripts.detect_platforms:main"
rentcafe-creds = "scripts.extract_rentcafe_credentials:main"
dev = "scripts.dev_bootstrap:main"
validate-building = "moxie.sync.push_availability:main"
create-admin = "scripts.create_admin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moxie", "scripts"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "pytest-httpx>=0.35.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
